tankify_backend/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
boto3==1.35.57
botocore==1.35.57
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
DateTime==5.5
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
uuid==1.30
Werkzeug==3.0.4
zope.interface==7.1.1
